fix: rename module according to review
