+++ cli/src/compiler.rs
